Debug hooks work, finished code to apply patches.
Unexpected behavior exists where patched bytes disappear if not applied to binary when debugger starts.

To get around this I will need to write a background thread to monitor the patched bytes database every so often and update a cached version which can be applied after the database has been patched.

TODO -- Complete the above.